moved guard clause code into seperate package
